Seperate the data load from streamlit.
